==> 08_Modelo_multinomial.qmd <==
```{r setup, include=FALSE, message=FALSE, error=FALSE, warning=FALSE}
knitr::opts_chunk$set(
  echo = TRUE,
  message = FALSE,
  warning = FALSE,
  cache = TRUE)
library(kableExtra)
library(tidyverse)
library(magrittr)
rstan::rstan_options(auto_write = TRUE) # speed up running time 
# Clearing the R environment by removing all variables from memory.
rm(list = ls())
tba <- function(dat, cap = NA){
  kable(dat,
      format = "html", digits =  4,
      caption = cap) %>% 
     kable_styling(bootstrap_options = "striped", full_width = F)%>%
         kable_classic(full_width = F, html_font = "Arial Narrow")
}
```


# Modelo multinomial para estimar personas por rango de edad y sexo

En el paso final, modelamos la estructura demográfica de las Unidades Geográficas Pequeñas a nivel de provincia. Esta estructura consta de 40 subgrupos basados en la combinación de género y grupos de edad. Para estimar la demografía de hogares no observados, utilizamos la predicción condicional basada en los resultados del paso anterior.

Dada la naturaleza del fenómeno en estudio, se considera adecuada una distribución multinomial para el recuento de individuos en los 40 grupos. Aquí tienes una explicación del modelo:

- $G_i$ sigue una distribución Multinomial $Multinomial(p_{G_i})$.

- Las log-odds de $p_{G_i}$ están modeladas por $x_i^T\beta$.

En este contexto, $G_i$ representa el recuento de cada uno de los cuarenta grupos demográficos. $p_{G_i}$ es un vector de tamaño 40 que contiene las probabilidades de que una persona clasificada en cada grupo resida en un hogar.


- Limpieza del entorno de R

El código elimina todos los objetos del entorno de R actual, asegurando un punto de partida limpio para las operaciones posteriores.

```{r}
# Loading necessary libraries for data analysis.
library(tidyverse)   # Data manipulation and visualization
library(data.table)  # Fast data manipulation
library(openxlsx)    # Reading Excel files
library(magrittr)    # Pipe operator
library(lme4)        # For fitting linear mixed-effects models
library(nnet)        # For fitting neural networks
cat("\f")  # Clears console output
```

-   Leyendo datos

```{r}
# Reading the preprocessed census data file containing information about households.
censo_vivienda <- readRDS("Recursos/06_Model_Multinomial/Data/02_censo_vivienda_personas.rds") 

# Filtering the census data to retain relevant columns for age and gender analysis.
censo_vivienda_age_gender <- censo_vivienda %>% 
  filter( !greenpoint2 %in% c("Sin informacion pero  n>0", 
                              "Sin informacion pero n>=0")) %>% 
  dplyr::select(PROV_ID, HOMBRES_GRUPO1_sum:MUJERES_GRUPO20_sum)
```

-   Preparación de datos:

    El código crea un nuevo conjunto de datos, censo_vivienda_edad_sexo, filtrando entradas específicas de censo_vivienda. Se seleccionan columnas relacionadas con grupos de edad y género y provincias. Luego, este conjunto de datos filtrado se agrega a nivel de PROV_ID utilizando la función summarise_if.
    

```{r}
# Summarizing the age and gender data by grouping it based on province (PROV_ID).
censo_personas <- censo_vivienda_age_gender %>% 
  group_by(PROV_ID) %>% 
  summarise_if(is.numeric, sum)
```

-   Modelo Multinomial:

    Se crea un modelo multinomial utilizando la función `multinom`. Este modelo predice la distribución de grupos de edad y género dentro de los hogares según la provincia (`PROV_ID`). El modelo se almacena en la variable `model`.

```{r}
# Fitting a multinomial model to estimate the distribution of age and gender within households,
# using province as the predictor.
model <- multinom( censo_personas[,-1] %>% as.matrix() ~ censo_personas$PROV_ID)
```

- Predicción del Modelo:

    La función `predict` se utiliza para predecir las probabilidades de distribución para el modelo multinomial. Los resultados de la predicción no se muestran aquí, pero se pueden obtener utilizando la función `predict`.
    

```{r, eval=FALSE}
# Predicting the distribution probabilities using the fitted model.
predict(model,type = "probs")
```
- Guardar el Modelo

    El modelo multinomial entrenado se guarda como un archivo RDS ("Recursos/06_Model_Multinomial/Data/Multinomial_model.rds") utilizando la función `saveRDS`.

```{r, eval=FALSE}
# Saving the fitted multinomial model to a file.
saveRDS(model, "Recursos/06_Model_Multinomial/Data/Multinomial_model.rds")
```
